Move crashtracking to its own module
