+++ crashtracker/src/constants.rs
