+++ tools/docker/Dockerfile.build
@@ -72,6 +72,7 @@ RUN cargo search nothing --limit 1
 
 # create stubs to cache compilation of dependendencies
 COPY [ "Cargo.lock", "Cargo.toml", "./"]
+COPY "crashtracker/Cargo.toml" "crashtracker/" 
 COPY "ddcommon/Cargo.toml" "ddcommon/" 
 COPY "ddcommon-ffi/Cargo.toml" "ddcommon-ffi/" 
 COPY "ddtelemetry/Cargo.toml" "ddtelemetry/"
